internal: move SectionCapture to section_capture module
